Moved the ProcessRecordsInputMatcher to its own class
